Загальні правки та фічв
2 Cinemachine камери які слідкують за деревом та корнем
При натисненні кнопки "E" камера плавно змінюється до дерева або до корня. При наступному етапі, коли дерево проростає віддаляється і трохи камера.
Додані нові спрайти дерев.

+++ Assets/Scripts/Resources/ResourcesHandler.cs
@@ -4,10 +4,10 @@ using UnityEngine;
 
 public class ResourcesHandler : MonoBehaviour
 {
-    [SerializeField] private float sunEnergyTick = 5f;
+    [SerializeField] private int sunEnergyTick = 5;
 
     private void OnParticleCollision(GameObject other)
     {
-        GameManager.Instance.AddResources(sunEnergyTick, 0f);
+        GameManager.Instance.AddResources(sunEnergyTick, 0);
     }
 }
